Fix type conversion warnings
- Change int count to NSUInteger in PBWebChangesController.m:49
- Change int aColumnIndex to NSInteger with explicit cast in PBUnsortableTableHeader.m:17,23
- Add explicit int casts for NSInteger returns in PBGitDefaults.m:73,83

+++ Classes/PBUnsortableTableHeader.m
@@ -14,13 +14,13 @@
 - (void)mouseDown:(NSEvent *)theEvent
 {
 	NSPoint location = [self convertPoint:[[self window] mouseLocationOutsideOfEventStream] fromView:[[self window] contentView]];
-	int aColumnIndex = [self columnAtPoint:location];
+	NSInteger aColumnIndex = [self columnAtPoint:location];
 
 	// If the user pressed on another column, reset
 	if (aColumnIndex != columnIndex)
 	{
 		clickCount = 1;
-		columnIndex = aColumnIndex;
+		columnIndex = (int)aColumnIndex;
 		[super mouseDown:theEvent];
 		return;
 	}
